feat : 이미지 좋아요 저장 && 검색 및 스크롤 로드 오류 해결

+++ app/src/main/java/com/sungho/searchapp/fragment/SearchFragment.kt
@@ -7,6 +7,7 @@ import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.widget.Toast
 import androidx.databinding.DataBindingUtil
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
@@ -20,7 +21,7 @@ class SearchFragment : Fragment() {
     private lateinit var binding : FragmentSearchBinding
     private lateinit var model : MainViewModel
 
-    private val adapter by lazy{ SearchAdapter() }
+    var adapter = SearchAdapter()
 
     var isLoading = false
     var page = 1
@@ -51,7 +52,6 @@ class SearchFragment : Fragment() {
         binding.itemRecyclerView.adapter = adapter
 
         binding.searchButton.setOnClickListener {
-            adapter.data.clear()
             query = "${binding.searchEditText.text}"
             page = 1
             model.search(query,page)
@@ -67,7 +67,8 @@ class SearchFragment : Fragment() {
                 val itemTotalCount = recyclerView.adapter?.itemCount?.minus(1)
 
                 // 스크롤 끝
-                if (!isLoading && !binding.itemRecyclerView.canScrollVertically(1) && lastVisibleItemPosition == itemTotalCount) {
+                if (!isLoading && !binding.itemRecyclerView.canScrollVertically(1)
+                    && lastVisibleItemPosition == itemTotalCount && model.searchItemList.value?.size != 0) {
                     adapter.deleteLoading()
                     model.load(query,++page)
                     isLoading = true
@@ -78,9 +79,24 @@ class SearchFragment : Fragment() {
 
     private fun setObserver(){
         model.searchItemEvent.observe(this, Observer{
-            adapter.setList(model.searchItemList.value ?: arrayListOf(),page)
-            adapter.notifyItemRangeInserted((page - 1) * 21, 21)    // v10 i10 separator1
-            isLoading = false
+            // 검색 결과가 0개일때
+            val size = model.searchItemList.value?.size ?: 0
+            if(size == 0) {
+                Toast.makeText(requireContext(),"검색 결과가 없습니다.",Toast.LENGTH_SHORT).show()
+                return@Observer
+            }
+
+            if(it.peekContent() == "new"){
+                adapter = SearchAdapter()
+                adapter.setList(model.searchItemList.value ?: arrayListOf(),page)
+                binding.itemRecyclerView.adapter = adapter
+            }else if(it.peekContent() == "load"){
+                adapter.setList(model.searchItemList.value ?: arrayListOf(),page)
+                adapter.notifyItemRangeInserted((page - 1) * size + 1, size + 1)
+                isLoading = false
+            }else{
+                Toast.makeText(requireContext(),"검색을 실패했습니다.",Toast.LENGTH_SHORT).show()
+            }
         })
     }
 }
